test: add more test
