+++ src/services/billing-services/index.js
@@ -4,7 +4,7 @@ import { deletePaymentService } from './delete-payment-service'
 import { setAsDefaultPaymentService } from './set-as-default-payment-service'
 import { addCreditService } from './add-credit-service'
 import { createPaymentMethodService } from './create-payment-method-service'
-import { loadYourServicePlanService } from './your-service-plan-service'
+import { loadYourServicePlanService } from './load-your-service-plan-service'
 
 export {
   listPaymentMethodsService,
